v1.2.7 Update
Hello, in this update I have worked on fixing a reported bug and adding
some user requested features.

Bug Fixes:
  - Random prize option was broke and would give multiple prizes even
when the bulk is turned off.

Added:
  - /Envoy ignore alias for the /Envoy stfu
  - New crate collects cooldown option. This allows you to add a bit
more PVP so players that have a cooldown fight for crates.
      You can bypass the cooldown with the envoy.bypass permission.

Changes:
  - /Envoy ignore now is the main command in the /envoy help instead of
/Envoy stfu.
  - envoy.stfu permission was changed to envoy.ignore.
  - Recoded how the message.yml works. The file looks the same but you
are now allowed to use a list to use multiple lines of messages.
  - Default messages.yml now has tips at the top to explain stuff.
  - Fixed method names not spelling Envoy right.

Removed:
  - Messages now do not automatically have a prefix. If you wish to
have a prefix add the %prefix% placeholder in the message.